+++ CHANGES.md
@@ -26,6 +26,9 @@ To be released.
      -  Removed `ITrie.Secure` property.
      -  Removed `secure` parameter from `MerkleTrie()`.
      -  Removed `secure` parameter from `PathCursor()`.
+ -  (Libplanet.Store) Added `ITrie.Diff()` method.  [[#3410]]
+ -  (Libplanet.Store) Removed `MerkleTrieExtensions.DifferentNodes()`
+    static method.  [[#3410]]
 
 ### Backward-incompatible network protocol changes
 
@@ -71,6 +74,7 @@ To be released.
 [#3402]: https://github.com/planetarium/libplanet/pull/3402
 [#3405]: https://github.com/planetarium/libplanet/pull/3405
 [#3407]: https://github.com/planetarium/libplanet/pull/3407
+[#3410]: https://github.com/planetarium/libplanet/pull/3410
 [RocksDB Read Only]: https://github.com/facebook/rocksdb/wiki/Read-only-and-Secondary-instances